requirements.txt
gradio>=4.0
requests
Pillow
numpy
opencv-python
torch
diffusers
transformers
accelerate
ultralytics
